pyproject.toml
[project]
name = "nass-mcp"
version = "0.1.0"
description = "MCP server for the USDA NASS API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.1",
    "python-dotenv>=1.1.1",
]
